perf: reduce Span._finish_ns overhead in tracer and profiling
  - Remove Hooks class usage from BaseContextProvider and Tracer
  - Replace hook-based callbacks with core event dispatcher system
  - Optimize _update_active logic to reduce redundant finished() checks
  - Cache span._local_root property lookup in profiling link_span
  - Streamline span activation flow in DefaultContextProvider
  - Convert SimpleMovingAverage to Cython for C-speed mathematical operations
  - Add encoder caching to reduce redundant _trace_id_64bits property lookups

Testing against a locally running version of PyPA warehouse with a single gunicorn worker with tracing and profiling enabled we saw an improvement of:

- a 4.9% improvement in requests per second (20.08 → 21.06 RPS)
- 4.6% reduction in response time (49.79ms → 47.48ms)
- elimination of 44,028 function calls (-7.2%)
