+++ ddtrace/_trace/provider.py
@@ -5,9 +5,9 @@ from typing import Callable
 from typing import Optional
 from typing import Union
 
-from ddtrace._hooks import Hooks
 from ddtrace._trace.context import Context
 from ddtrace._trace.span import Span
+from ddtrace.internal import core
 from ddtrace.internal.logger import get_logger
 
 
@@ -30,46 +30,17 @@ class BaseContextProvider(metaclass=abc.ABCMeta):
     * the ``activate`` method, that sets the current active ``Context``
     """
 
-    def __init__(self) -> None:
-        self._hooks = Hooks()
-
     @abc.abstractmethod
     def _has_active_context(self) -> bool:
         pass
 
     @abc.abstractmethod
     def activate(self, ctx: Optional[ActiveTrace]) -> None:
-        self._hooks.emit(self.activate, ctx)
+        core.dispatch("ddtrace.context_provider.activate", (ctx,))
 
     @abc.abstractmethod
     def active(self) -> Optional[ActiveTrace]:
         pass
-
-    def _on_activate(
-        self, func: Callable[[Optional[Union[Span, Context]]], Any]
-    ) -> Callable[[Optional[Union[Span, Context]]], Any]:
-        """Register a function to execute when a span is activated.
-
-        Can be used as a decorator.
-
-        :param func: The function to call when a span is activated.
-                     The activated span will be passed as argument.
-        """
-        self._hooks.register(self.activate, func)
-        return func
-
-    def _deregister_on_activate(
-        self, func: Callable[[Optional[Union[Span, Context]]], Any]
-    ) -> Callable[[Optional[Union[Span, Context]]], Any]:
-        """Unregister a function registered to execute when a span is activated.
-
-        Can be used as a decorator.
-
-        :param func: The function to stop calling when a span is activated.
-        """
-
-        self._hooks.deregister(self.activate, func)
-        return func
 
     def __call__(self, *args: Any, **kwargs: Any) -> Optional[ActiveTrace]:
         """Method available for backward-compatibility. It proxies the call to
@@ -111,13 +82,13 @@ class DefaultContextProvider(BaseContextProvider):
         When a span finishes, the active span becomes its parent.
         If no parent exists and the context is reactivatable, that context is restored.
         """
-        if span.finished:
-            new_active: Optional[Span] = span
-            while new_active and new_active.finished:
-                if new_active._parent is None and new_active._parent_context and new_active._parent_context._reactivate:
-                    self.activate(new_active._parent_context)
-                    return new_active._parent_context
-                new_active = new_active._parent
+        new_active: Optional[Span] = span
+        # PERF: Avoid calling `Span.finished` more than once per span. This is a computed property.
+        while new_active and new_active.finished:
+            if new_active._parent is None and new_active._parent_context and new_active._parent_context._reactivate:
+                self.activate(new_active._parent_context)
+                return new_active._parent_context
+            new_active = new_active._parent
+        if new_active is not span:
             self.activate(new_active)
-            return new_active
-        return span
+        return new_active
