+++ ddtrace/internal/_encoding.pyx
@@ -491,6 +491,7 @@ cdef class MsgpackEncoderBase(BufferedEncoder):
         cdef int ret
         cdef Py_ssize_t L
         cdef void * dd_origin = NULL
+        cdef unsigned long long trace_id_64bits = 0
 
         L = len(trace)
         if L > ITEM_LIMIT:
@@ -503,9 +504,12 @@ cdef class MsgpackEncoderBase(BufferedEncoder):
         if L > 0 and trace[0].context is not None and trace[0].context.dd_origin is not None:
             dd_origin = self.get_dd_origin_ref(trace[0].context.dd_origin)
 
+
+        # PERF: _trace_id_64bits is a computed property, cache/convert once for all spans
+        trace_id_64bits = trace[0]._trace_id_64bits
         for span in trace:
             try:
-                ret = self.pack_span(span, dd_origin)
+                ret = self.pack_span(span, trace_id_64bits, dd_origin)
             except Exception as e:
                 raise RuntimeError("failed to pack span: {!r}. Exception: {}".format(span, e))
 
@@ -555,7 +559,7 @@ cdef class MsgpackEncoderBase(BufferedEncoder):
     cpdef flush(self):
         raise NotImplementedError()
 
-    cdef int pack_span(self, object span, void *dd_origin) except? -1:
+    cdef int pack_span(self, object span, unsigned long long trace_id_64bits, void *dd_origin) except? -1:
         raise NotImplementedError()
 
 
@@ -737,7 +741,7 @@ cdef class MsgpackEncoderV04(MsgpackEncoderBase):
 
         raise TypeError("Unhandled metrics type: %r" % type(metrics))
 
-    cdef int pack_span(self, object span, void *dd_origin) except? -1:
+    cdef int pack_span(self, object span, unsigned long long trace_id_64bits, void *dd_origin) except? -1:
         cdef int ret
         cdef Py_ssize_t L
         cdef int has_span_type
@@ -768,7 +772,7 @@ cdef class MsgpackEncoderV04(MsgpackEncoderBase):
             ret = pack_bytes(&self.pk, <char *> b"trace_id", 8)
             if ret != 0:
                 return ret
-            ret = pack_number(&self.pk, span._trace_id_64bits)
+            ret = msgpack_pack_unsigned_long_long(&self.pk, trace_id_64bits)
             if ret != 0:
                 return ret
 
@@ -1013,7 +1017,7 @@ cdef class MsgpackEncoderV05(MsgpackEncoderBase):
     cdef void * get_dd_origin_ref(self, str dd_origin):
         return <void *> PyLong_AsLong(self._st._index(dd_origin))
 
-    cdef int pack_span(self, object span, void *dd_origin) except? -1:
+    cdef int pack_span(self, object span, unsigned long long trace_id_64bits, void *dd_origin) except? -1:
         cdef int ret
 
         ret = msgpack_pack_array(&self.pk, 12)
@@ -1030,8 +1034,7 @@ cdef class MsgpackEncoderV05(MsgpackEncoderBase):
         if ret != 0:
             return ret
 
-        _ = span._trace_id_64bits
-        ret = msgpack_pack_uint64(&self.pk, _ if _ is not None else 0)
+        ret = msgpack_pack_unsigned_long_long(&self.pk, trace_id_64bits)
         if ret != 0:
             return ret
