+++ setup.py
@@ -796,6 +796,11 @@ setup(
                     sources=["ddtrace/internal/_tagset.pyx"],
                     language="c",
                 ),
+                Cython.Distutils.Extension(
+                    "ddtrace.internal.sma",
+                    sources=["ddtrace/internal/sma.pyx"],
+                    language="c",
+                ),
                 Extension(
                     "ddtrace.internal._encoding",
                     ["ddtrace/internal/_encoding.pyx"],
